+++ components/ms-business-central/src/actions/items/list.ts
@@ -10,21 +10,7 @@ export const listItems = action({
   },
   perform: async (
     context,
-    {
-      companyId,
-      connection,
-      fetchAll,
-      $orderBy,
-      $format,
-      $expand,
-      $count,
-      $filter,
-      $top,
-      $skipToken,
-      $skip,
-      $search,
-      $select,
-    },
+    { companyId, connection, fetchAll, odataQueryParams },
   ) => {
     const client = getMsBusinessCentralClient(
       connection,
@@ -32,22 +18,22 @@ export const listItems = action({
       context.debug.enabled,
     );
     const params = {
-      $orderBy,
-      $format,
-      $expand,
-      $count,
-      $filter,
-      $skipToken,
-      $skip,
-      $search,
-      $select,
+      $orderBy: odataQueryParams.$orderBy,
+      $format: odataQueryParams.$format,
+      $expand: odataQueryParams.$expand,
+      $count: odataQueryParams.$count,
+      $filter: odataQueryParams.$filter,
+      $skipToken: odataQueryParams.$skipToken,
+      $skip: odataQueryParams.$skip,
+      $search: odataQueryParams.$search,
+      $select: odataQueryParams.$select,
     };
     return await paginateResults({
       client,
       endpoint: `/companies(${companyId})/items`,
       params,
       fetchAll,
-      pageSize: $top,
+      pageSize: odataQueryParams.$top,
     });
   },
   inputs,