+++ components/ms-business-central/src/inputs/purchaseOrderLines/listPurchaseOrderLinesInputs.ts
@@ -1,10 +1,10 @@
 import { companyId } from "../accounts/getAccountsInputs";
-import { connectionInput, fetchAll, odataParams } from "../general";
+import { connectionInput, fetchAll, odataQueryParams } from "../general";
 import { purchaseOrderId } from "../purchaseOrders/shared";
 export const listPurchaseOrderLinesInputs = {
   connection: connectionInput,
   companyId,
   purchaseOrderId,
   fetchAll,
-  ...odataParams,
+  odataQueryParams,
 };